backend/requirements.txt
connexion[swagger-ui]
psycopg2
python-jose[cryptography]
six>=1.9
flask_cors